fixed fejlkode i wishlistrepo. Fixed noget getmapping

+++ src/main/java/org/example/projekt4_gruppe_3/Controller/WishlistController.java
@@ -1,5 +1,6 @@
 package org.example.projekt4_gruppe_3.Controller;
 
+import jakarta.servlet.http.HttpSession;
 import org.example.projekt4_gruppe_3.Model.User;
 import org.example.projekt4_gruppe_3.Model.Wishlist;
 import org.example.projekt4_gruppe_3.Repository.UserRepository;
@@ -21,29 +22,63 @@ public class WishlistController {
     @Autowired
     WishlistRepository wishlistRepo;
 
+    @Autowired
+    UserRepository userRepo;
+
     @GetMapping("/getAllWishlists")
-    public String getAllWishlists(Model model) {
-        ArrayList<Wishlist> wishlists = wishlistRepo.getAllWishlists();
-        model.addAttribute("Wishlists", wishlists);
+    public String getAllWishlists(Model model, HttpSession session) {
+
+        if (!isUserLoggedIn(session)){
+            return "redirect:/login";
+        }
+
+        ArrayList<Wishlist> wishlists = new ArrayList<>();
+
+        wishlists = wishlistRepo.getAllWishlists();
+
+
+        User loggedInUser = (User) session.getAttribute("loggedInUser");
+        model.addAttribute("username", loggedInUser.getFullName());
+
+        model.addAttribute("wishlists", wishlists);
 
         return "Profile";
     }
 
+    boolean isUserLoggedIn(HttpSession session) {
+        return session.getAttribute("loggedInUser") != null;
+    }
+
     @GetMapping("/getWishlist")
-    public String getWishlist(@RequestParam("id") int id, Model model) {
+    public String getWishlist(@RequestParam("id") int id, Model model, HttpSession session) {
+
+        if (!isUserLoggedIn(session)){
+            return "redirect:/login";
+        }
+
         Wishlist wishlist = wishlistRepo.getWishlistById(id);
         model.addAttribute("wishlist", wishlist);
         return "showWishlist";
     }
 
     @PostMapping("/deleteWishlist")
-    public String deleteWishlist(@RequestParam("id") int id) {
+    public String deleteWishlist(@RequestParam("id") int id, HttpSession session) {
+
+        if (!isUserLoggedIn(session)){
+            return "redirect:/login";
+        }
+
         wishlistRepo.deleteWishlistById(id);
         return "redirect:/";
     }
 
     @GetMapping("getUpdateWishlist")
-    public String getUpdateWishlist(@RequestParam("id") int id, Model model) {
+    public String getUpdateWishlist(@RequestParam("id") int id, Model model, HttpSession session) {
+
+        if (!isUserLoggedIn(session)){
+            return "redirect:/login";
+        }
+
         Wishlist wishlist = wishlistRepo.getWishlistById(id);
         model.addAttribute("wishlist", wishlist);
         return "updateWishlist";
